+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - `rp_thread_logging = False` config parameter handling, by @HardNorth
+- Recursive thread init issue in case of `rp_thread_logging = True`, by @HardNorth
 ### Changed
 - Client version updated on [5.3.2](https://github.com/reportportal/client-Python/releases/tag/5.3.2), by @HardNorth